requirements.txt
SQLAlchemy
PyYAML
Faker
numpy